+++ CHANGELOG.rst
@@ -9,6 +9,7 @@ CHANGELOG
 * feature: Add timestamp to secondary status in training job output
 * bug-fix: Local Mode: Set correct default values for additional_volumes and additional_env_vars
 * enhancement: Local Mode: support nvidia-docker2 natively
+* warning: Frameworks: add warning for upcoming breaking change that makes framework_version required
 
 1.11.2
 ======